feat: api와 통신 및 recoil 설정
