+++ project/package.json
@@ -4,6 +4,7 @@
   "private": true,
   "dependencies": {
     "@actbase/react-daum-postcode": "^1.0.4",
+    "@react-oauth/google": "^0.6.0",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
@@ -18,10 +19,13 @@
     "react-daum-postcode": "^3.1.1",
     "react-dom": "^18.2.0",
     "react-geocode": "^0.2.3",
+    "react-google-login": "^5.2.2",
     "react-icons": "^4.7.1",
     "react-modal": "^3.16.1",
     "react-router-dom": "^6.6.2",
     "react-scripts": "5.0.1",
+    "recoil": "^0.7.6",
+    "recoil-persist": "^4.2.0",
     "typescript": "^4.9.4",
     "web-vitals": "^2.1.4"
   },
@@ -52,5 +56,6 @@
   "devDependencies": {
     "@types/styled-components": "^5.1.26",
     "styled-components": "^5.3.6"
-  }
+  },
+  "proxy": "http://pikcha36.o-r.kr:8080"
 }
